Add DialogProvider

+++ components/Tournaments/TournamentSettings.tsx
@@ -12,13 +12,28 @@ import { TournamentResponse } from '@/prisma/prisma-types';
 import { useParams, useRouter } from 'next/navigation';
 import useSWR from 'swr';
 import { ClientSideApi } from '@/services/ClientSideApi';
+import { useDialog } from '@/contexts/DialogProvider';
 
 export const TournamentSettings = () => {
     const router = useRouter();
+    const { confirmDialog } = useDialog()
     const { slug } = useParams<{ slug: string }>();
     const { data } = useSWR<TournamentResponse>(`/tournaments/${slug}`);
 
+    if (!data) {
+        return null;
+    }
+
     const handleDelete = async () => {
+        const confirm = await confirmDialog({
+            title: `Delete tournament ${data.name}`,
+            description: 'Are you sure you want to delete this tournament?',
+        })
+
+        if (!confirm) {
+            return
+        }
+
         const deletedTournament = await ClientSideApi.deleteTournament(slug);
 
         if (!deletedTournament.id) {
@@ -27,10 +42,6 @@ export const TournamentSettings = () => {
 
         router.replace('/tournaments');
     };
-
-    if (!data) {
-        return null;
-    }
 
     return (
         <DropdownMenu>
